vikings-frontend/API/requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.3
plotly==4.5.4
psycopg2-binary==2.8.4
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
waitress==1.4.3
Werkzeug==1.0.0
